refactor: avoid barrel import
