+++ apps/main/src/dex/components/PagePool/Withdraw/components/FormWithdraw.tsx
@@ -23,7 +23,7 @@ import { CurveApi, Pool, PoolData } from '@/dex/types/main.types'
 import { notify } from '@evm-ui/features/connect-wallet'
 import { useTokenUsdRates } from '@evm-ui/lib/model/entities/token-usd-rate'
 import { FormContent } from '@evm-ui/widgets/DetailPageLayout/FormContent'
-import { EvmSlippageToleranceActionInfo } from '@evm-ui/widgets/SlippageSettings'
+import { EvmSlippageToleranceActionInfo } from '@evm-ui/widgets/SlippageSettings/EvmSlippageToleranceActionInfo'
 import { Box } from '@legacy-ui/Box'
 import { Checkbox } from '@legacy-ui/Checkbox'
 import { Radio, RadioGroup } from '@legacy-ui/Radio'
