+++ apps/main/src/dex/components/PageRouterSwap/index.tsx
@@ -34,7 +34,8 @@ import { ActionInfo, ActionInfoGasEstimate } from '@evm-ui/shared/ui/ActionInfo'
 import { LargeTokenInput } from '@evm-ui/shared/ui/LargeTokenInput'
 import { decimal, formatNumber } from '@evm-ui/utils'
 import { getPriceImpactDisplay } from '@evm-ui/widgets/DetailPageLayout/price-impact.util'
-import { EvmSlippageToleranceActionInfo, type SlippageType } from '@evm-ui/widgets/SlippageSettings'
+import { EvmSlippageToleranceActionInfo } from '@evm-ui/widgets/SlippageSettings/EvmSlippageToleranceActionInfo'
+import { type SlippageType } from '@evm-ui/widgets/SlippageSettings/slippage.utils'
 import { AlertBox } from '@legacy-ui/AlertBox'
 import { Icon } from '@legacy-ui/Icon'
 import { IconButton } from '@legacy-ui/IconButton'
