+++ packages/evm-ui/src/shared/ui/Chart/index.ts
@@ -12,4 +12,3 @@ export * from './SelectTimeOption'
 export * from './ToggleBandsChartButton'
 export * from './hooks/useEChartsTooltip'
 export * from './chart.utils'
-export { ChartStateWrapper } from '@evm-ui/shared/ui/Chart/ChartStateWrapper'
