+++ packages/evm-ui/src/widgets/SlippageSettings/index.tsx
@@ -1,2 +1,1 @@
-export { EvmSlippageToleranceActionInfo } from './EvmSlippageToleranceActionInfo'
 export { type SlippageType } from './slippage.utils'
